feat: implementa ClienteFisicoService

+++ src/main/java/com/gestaooserp/dev/repository/ClienteFisicoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.ClienteFisico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteFisicoRepository extends JpaRepository<ClienteFisico,Integer> {
+public interface ClienteFisicoRepository extends JpaRepository<ClienteFisico,Long> {
 }
